Add initial top N
