+++ web-common/src/features/canvas/components/charts/types.ts
@@ -9,6 +9,7 @@ export interface FieldConfig {
   type: "quantitative" | "ordinal" | "nominal" | "temporal";
   timeUnit?: string; // For temporal fields
   sort?: ChartSortDirection;
+  topN?: number;
 }
 
 export interface ChartConfig {
